fixup! fix(ivy): ngcc should compile entry-points in the correct order
