+++ packages/compiler-cli/src/ngcc/src/packages/dependency_host.ts
@@ -7,114 +7,13 @@
  */
 
 import * as path from 'canonical-path';
-import {DepGraph} from 'dependency-graph';
 import * as fs from 'fs';
 import * as ts from 'typescript';
-
-import {EntryPoint} from './entry_point';
-
-
-/**
- * Holds information about entry points that were ignored because
- * they have depedencies that are missing.
- *
- * This might not be an error, because the entry point might not actually be used
- * in the application. If it is used then the `ngc` application compilation would
- * fail also.
- *
- * For example, an application use the `@angular/router` package. This package includes an
- * entry-point called `@angular/router/upgrade`, which has a dependency on the
- * `@angular/upgrade` package. If the application never uses code from `@angular/router/upgrade`
- * then there is no need for `@angular/upgrade` to be installed.
- *
- * In this case the ngcc tool should just ignore the `@angular/router/upgrade` end-point.
- */
-export interface IgnoredEntryPoint {
-  entryPoint: EntryPoint;
-  missingDeps: string[];
-}
-
-/**
- * Holds information about dependencies of an entry-point that do not need to be processed
- * by the ngcc tool.
- *
- * For example, the `rxjs` package does not contain any Angular decorators that need to be
- * compiled and so this can be safely ignored by ngcc.
- */
-export interface IgnoredDependency {
-  entryPoint: EntryPoint;
-  dependencyPath: string;
-}
-
-/**
- * The result of sorting the entry-points by their dependencies.
- *
- * The `entryPoints` array will be ordered so that no entry point depends upon an entry point that
- * appears later in the array.
- *
- * Some entry points or their dependencies may be have been ignored. These are captured for
- * diagnostic purposes in `ignoredEntryPoints` and `ignoredDependencies` respectively.
- */
-export interface SortedEntryPointsInfo {
-  entryPoints: EntryPoint[];
-  ignoredEntryPoints: IgnoredEntryPoint[];
-  ignoredDependencies: IgnoredDependency[];
-}
-
-/**
- * Sort the array of entry points so that the dependant entry points always come later than
- * their dependencies in the array.
- * @param entryPoints An array entry points to sort.
- * @returns the result of sorting the entry points.
- */
-export function sortEntryPointsByDependency(entryPoints: EntryPoint[]): SortedEntryPointsInfo {
-  const ignoredEntryPoints: IgnoredEntryPoint[] = [];
-  const ignoredDependencies: IgnoredDependency[] = [];
-  const graph = new DepGraph();
-
-  // Add the entry ponts to the graph as nodes
-  entryPoints.forEach(entryPoint => graph.addNode(entryPoint.path, entryPoint));
-
-  // Now add the dependencies between them
-  entryPoints.forEach(entryPoint => {
-    const entryPointPath = entryPoint.esm2015;
-    if (!entryPointPath) {
-      throw new Error(`Esm2015 format missing in '${entryPoint.path}' entry-point.`);
-    }
-
-    const dependencies = new Set<string>();
-    const missing = new Set<string>();
-    Helpers.computeDependencies(entryPointPath, dependencies, missing);
-
-    if (missing.size > 0) {
-      const nodesToRemove = [entryPoint.path, ...graph.dependantsOf(entryPoint.path)];
-      nodesToRemove.forEach(node => {
-        ignoredEntryPoints.push(
-            {entryPoint: graph.getNodeData(node) as EntryPoint, missingDeps: Array.from(missing)});
-        graph.removeNode(node);
-      });
-    } else {
-      dependencies.forEach(dependencyPath => {
-        if (graph.hasNode(dependencyPath)) {
-          graph.addDependency(entryPoint.path, dependencyPath);
-        } else {
-          ignoredDependencies.push({entryPoint, dependencyPath});
-        }
-      });
-    }
-  });
-
-  return {
-    entryPoints: graph.overallOrder().map(path => graph.getNodeData(path) as EntryPoint),
-    ignoredEntryPoints,
-    ignoredDependencies
-  };
-}
 
 /**
  * Helper functions for computing dependencies.
  */
-export class Helpers {
+export class DependencyHost {
   /**
    * Get a list of the resolved paths to all the dependencies of this entry point.
    * @param from An absolute path to the file whose dependencies we want to get.
@@ -127,11 +26,11 @@ export class Helpers {
    * array of
    * import specifiers for dependencies that were `missing`.
    */
-  static computeDependencies(
+  computeDependencies(
       from: string, resolved: Set<string>, missing: Set<string>,
       internal: Set<string> = new Set()): void {
     const fromContents = fs.readFileSync(from, 'utf8');
-    if (!Helpers.hasImportOrReeportStatements(fromContents)) {
+    if (!this.hasImportOrReeportStatements(fromContents)) {
       return;
     }
 
@@ -140,21 +39,21 @@ export class Helpers {
         ts.createSourceFile(from, fromContents, ts.ScriptTarget.ES2015, false, ts.ScriptKind.JS);
     sf.statements
         // filter out statements that are not imports or reexports
-        .filter(Helpers.isStringImportOrReexport)
+        .filter(this.isStringImportOrReexport)
         // Grab the id of the module that is being imported
         .map(stmt => stmt.moduleSpecifier.text)
         // Resolve this module id into an absolute path
         .forEach(importPath => {
           if (importPath.startsWith('.')) {
             // This is an internal import so follow it
-            const internalDependency = Helpers.resolveInternal(from, importPath);
+            const internalDependency = this.resolveInternal(from, importPath);
             // Avoid circular dependencies
             if (!internal.has(internalDependency)) {
               internal.add(internalDependency);
-              Helpers.computeDependencies(internalDependency, resolved, missing, internal);
+              this.computeDependencies(internalDependency, resolved, missing, internal);
             }
           } else {
-            const externalDependency = Helpers.tryResolveExternal(from, importPath);
+            const externalDependency = this.tryResolveExternal(from, importPath);
             if (externalDependency !== null) {
               resolved.add(externalDependency);
             } else {
@@ -170,7 +69,7 @@ export class Helpers {
    * @param to the module specifier of the internal dependency to resolve
    * @returns the resolved path to the import.
    */
-  static resolveInternal(from: string, to: string): string {
+  resolveInternal(from: string, to: string): string {
     const fromDirectory = path.dirname(from);
     // `fromDirectory` is absolute so we don't need to worry about telling `require.resolve`
     // about it - unlike `tryResolve` below.
@@ -194,8 +93,8 @@ export class Helpers {
    * @returns the resolved path to the entry point directory of the import or null
    * if it cannot be resolved.
    */
-  static tryResolveExternal(from: string, to: string): string|null {
-    const externalDependency = Helpers.tryResolve(from, `${to}/package.json`);
+  tryResolveExternal(from: string, to: string): string|null {
+    const externalDependency = this.tryResolve(from, `${to}/package.json`);
     return externalDependency && path.dirname(externalDependency);
   }
 
@@ -207,7 +106,7 @@ export class Helpers {
    * @returns an absolute path to the entry-point of the dependency or null if it could not be
    * resolved.
    */
-  static tryResolve(from: string, to: string): string|null {
+  tryResolve(from: string, to: string): string|null {
     try {
       return require.resolve(to, {paths: [from]});
     } catch (e) {
@@ -220,7 +119,7 @@ export class Helpers {
    * @param stmt the statement node to check.
    * @returns true if the statement is an import with a string literal module specifier.
    */
-  static isStringImportOrReexport(stmt: ts.Statement): stmt is ts.ImportDeclaration&
+  isStringImportOrReexport(stmt: ts.Statement): stmt is ts.ImportDeclaration&
       {moduleSpecifier: ts.StringLiteral} {
     return ts.isImportDeclaration(stmt) ||
         ts.isExportDeclaration(stmt) && !!stmt.moduleSpecifier &&
@@ -237,7 +136,7 @@ export class Helpers {
    * @returns false if there are definitely no import or re-export statements
    * in this file, true otherwise.
    */
-  static hasImportOrReeportStatements(source: string): boolean {
+  hasImportOrReeportStatements(source: string): boolean {
     return /(import|export)\s+[^\n]+from/.test(source);
   }
 }
